lua: ensure directories have trailing slash
Currently, when autocompleting a file that is a directory, the text that
is actually completed in the cmdline has no trailing slash. This makes
it annoying to type file paths as the text in the label does not match
the text that is actually inserted. This commit fixes the behavior by
ensuring that inserted text always has a trailing slash.

Note: although this achieves the desired results, I'm not sure if this
approach is semantically correct; please advise.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>